sf_mobile_base: get version from sf.app
