+++ shopfloor_mobile_base/__init__.py
@@ -1,1 +1,2 @@
 from . import controllers
+from . import models
