+++ shopfloor_mobile_base/models/__init__.py
@@ -1,0 +1,1 @@
+from . import shopfloor_app
